requirements.txt
PyQt6>=6.5.0
PyYAML>=6.0
rapidfuzz>=3.5.0
pandas